Adding validation, user roles and e-mail forgot password functionality

+++ app/Http/Controllers/CommentController.php
@@ -38,6 +38,11 @@ class CommentController extends Controller
     public function store(Request $request)
     {
         $forumId = Forum::where('reservoir_id', $request->reservoirId)->get()->first()->id;
+
+        if (empty($forumId)) {
+            return redirect()->route('home')->with('error', 'Nevar pievienot komentāru neeksistējošai diskusijai. ');
+        }
+
         $userId = Auth::user()->getAuthIdentifier();
         $commentData = $request->commentData;
 
@@ -81,6 +86,14 @@ class CommentController extends Controller
      */
     public function update(Request $request, Comment $comment)
     {
+        if ($comment->user_id != Auth::id()) {
+            return redirect()->route('forum.show', $comment->forum_id)->with('error', 'Šis komentārs nav pievienots ar Jūsu lietotāju.');
+        }
+
+        if (!Comment::find($comment->id)) {
+            return redirect()->route('forum.show', $comment->forum_id)->with('error', 'Komentārs ar doto identifikatoru neeksistē.');
+        }
+
         $comment->content = $request->commentData;
         $comment->update();
     }
@@ -93,6 +106,10 @@ class CommentController extends Controller
      */
     public function destroy(Comment $comment)
     {
+        if(Auth::guest() || Auth::id() != $comment->user_id) {
+            return redirect()->route('forum.show', $comment->forum_id)->with('error', 'Neautorizēts vai nepareizs lietotājs centās izdzēst komentāru.');
+        }
+
         $comment->delete();
     }
 }
